+++ packages/shared/lib/participation/index.ts
@@ -3,5 +3,6 @@ export * as constants from './constants'
 export * as stores from './stores'
 export * as types from './types'
 
+export * from './account'
 export * from './participation'
 export * from './staking'
